requirements.txt
django~=4.2.3
djangorestframework~=3.14.0
django-rest-passwordreset
psycopg2-binary
pyyaml~=6.0
ujson~=5.8.0
requests~=2.31.0
pytest~=7.4.0